requirements.txt
pandas
multimethod
decord
tqdm
torch
torchcodec
torchvision
toolz
av
opencv-python
numpy
ray
python-dotenv
boto3
zmq
ultralytics
